apps/keycloak-config/requirements.txt
python-keycloak==4.7.2
requests
